Revert "Merge branch 'develop' into feature/rewards-cumulative"
This reverts commit 519ce23c740f0440c4b9dc3dbcd16e7107b5bd71, reversing
changes made to 25d48d2ec22e25c3fb75d5e807124aabbde9b2ca.
