+++ .github/workflows/ci.yml
@@ -326,7 +326,7 @@ jobs:
     needs: build-rococo
     runs-on: self-hosted
     timeout-minutes: 1440
-    if: ${{ (success() || github.event.inputs.skipBuild == 'true') && !contains(github.event.pull_request.labels.*.name, 'skip-e2e-tests') }}
+    if: "!contains(github.event.pull_request.labels.*.name, 'skip-e2e-tests')"
     env:
       API_URL: 'ws://127.0.0.1:8844'
       TEST_PALLET_ADDRESS: ${{ secrets.E2E_TEST_PALLET_ADDRESS }}
@@ -392,7 +392,7 @@ jobs:
         testFiles: ${{ fromJson(needs.build-matrix.outputs.testFiles_matrix) }}
 
     timeout-minutes: 1440
-    if: ${{ (success() || github.event.inputs.skipBuild == 'true') && !contains(github.event.pull_request.labels.*.name, 'skip-e2e-tests') }}
+    if: "!contains(github.event.pull_request.labels.*.name, 'skip-e2e-tests')"
     env:
       API_URL: 'ws://127.0.0.1:8844'
       TEST_PALLET_ADDRESS: ${{ secrets.E2E_TEST_PALLET_ADDRESS }}
@@ -529,7 +529,7 @@ jobs:
   e2e-test-sequential-self-hosted:
     needs: build-rococo
     runs-on: self-hosted
-    if: ${{ (success() || github.event.inputs.skipBuild == 'true') && !contains(github.event.pull_request.labels.*.name, 'skip-e2e-tests') }}
+    if: "!contains(github.event.pull_request.labels.*.name, 'skip-e2e-tests')"
     env:
       API_URL: 'ws://127.0.0.1:8844'
       TEST_PALLET_ADDRESS: ${{ secrets.E2E_TEST_PALLET_ADDRESS }}
