+++ pallets/bootstrap/src/migrations.rs
@@ -33,12 +33,12 @@ pub mod v1 {
 
 			log!(
 				info,
-				"Running V1 migration with current storage version {:?} / onchain {:?}",
+				"Running migration with current storage version {:?} / onchain {:?}",
 				current,
 				onchain
 			);
 
-			if current == 2 && onchain == 0 {
+			if current == 1 && onchain == 0 {
 				// this is safe to execute on any runtime that has a bounded number of pools.
 
 				if Phase::<T>::get() == BootstrapPhase::Finished && ActivePair::<T>::get().is_none()
@@ -47,7 +47,6 @@ pub mod v1 {
 					log!(info, "Filled ActivePair with default value");
 					T::DbWeight::get().reads_writes(3, 1)
 				} else {
-					log!(info, "No extra actions needed");
 					T::DbWeight::get().reads_writes(3, 0)
 				}
 			} else {
@@ -67,73 +66,6 @@ pub mod v1 {
 		fn post_upgrade() -> Result<(), &'static str> {
 			log!(info, "Bootstrap::post_upgrade");
 			assert_eq!(Pallet::<T>::on_chain_storage_version(), 1);
-			Ok(())
-		}
-	}
-}
-
-pub mod v2 {
-	use super::*;
-	use crate::log;
-	use frame_support::traits::OnRuntimeUpgrade;
-	use sp_runtime::traits::UniqueSaturatedInto;
-
-	pub struct MigrateToV2<T>(sp_std::marker::PhantomData<T>);
-	impl<T: Config> OnRuntimeUpgrade for MigrateToV2<T> {
-		fn on_runtime_upgrade() -> Weight {
-			let current = Pallet::<T>::current_storage_version();
-			let onchain = Pallet::<T>::on_chain_storage_version();
-
-			log!(
-				info,
-				"Running V2 migration with current storage version {:?} / onchain {:?}",
-				current,
-				onchain
-			);
-
-			if current == 2 && onchain < current {
-				if let Some((start, whitelist_length, public_length, _)) =
-					BootstrapSchedule::<T>::get()
-				{
-					let start_block: u32 = start.unique_saturated_into();
-					log!(info, "migrating vested provisions storage");
-					VestedProvisions::<T>::translate::<(Balance, BlockNrAsBalance), _>(
-						|_: T::AccountId,
-						 _: TokenId,
-						 (amount, lock_end): (Balance, BlockNrAsBalance)| {
-							Some((
-								amount,
-								(start_block + whitelist_length + public_length).into(),
-								lock_end,
-							))
-						},
-					);
-
-					T::DbWeight::get().reads_writes(
-						1,
-						(VestedProvisions::<T>::iter().count() as usize).try_into().unwrap(),
-					)
-				} else {
-					log!(info, "No ongoing bootstrap, migration not needed");
-					T::DbWeight::get().reads(1)
-				}
-			} else {
-				log!(info, "Migration did not executed. This probably should be removed");
-				T::DbWeight::get().reads(2)
-			}
-		}
-
-		#[cfg(feature = "try-runtime")]
-		fn pre_upgrade() -> Result<(), &'static str> {
-			log!(info, "Bootstrap::pre_upgrade");
-			assert_eq!(Pallet::<T>::on_chain_storage_version(), 2);
-			Ok(())
-		}
-
-		#[cfg(feature = "try-runtime")]
-		fn post_upgrade() -> Result<(), &'static str> {
-			log!(info, "Bootstrap::post_upgrade");
-			assert_eq!(Pallet::<T>::on_chain_storage_version(), 2);
 			Ok(())
 		}
 	}
