+++ primitives/multipurpose-liquidity/src/lib.rs
@@ -6,13 +6,13 @@ use sp_runtime::RuntimeDebug;
 #[derive(Eq, PartialEq, Clone, Encode, Decode, RuntimeDebug, TypeInfo)]
 pub enum ActivateKind {
 	AvailableBalance,
-	StakedUnactivatedReserves,
+	StakedUnactivatedLiquidity,
 	UnspentReserves,
 }
 
 #[derive(Eq, PartialEq, Clone, Encode, Decode, RuntimeDebug, TypeInfo)]
 pub enum BondKind {
 	AvailableBalance,
-	ActivatedUnstakedReserves,
+	ActivatedUnstakedLiquidity,
 	UnspentReserves,
 }
