+++ runtime/mangata-rococo/src/xcm_config.rs
@@ -201,7 +201,6 @@ pub type Trader = (
 	FixedRateOfFungible<KarPerSecond, ToTreasury>,
 	FixedRateOfFungible<KusdPerSecond, ToTreasury>,
 	FixedRateOfFungible<TurPerSecond, ToTreasury>,
-	FixedRateOfFungible<BncPerSecond, ToTreasury>,
 );
 
 pub struct XcmConfig;
